# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "MEDIA BASE URL : 3.17.254.50:4999/upload/\n\nHeader -> security_key : JHNlcnZpY2VAdmFsZXlvdUBwcm92aWRlciQ= \n\n Socket Url :  https://docs.google.com/document/d/1CEJFBiFLvaOq5uac9Hph5jB1RGSXycyk5JI76tnWHuY/edit?usp=sharing\n\n",
    "version": "1.0.0",
    "title": "Valeyou App"
  },
  "host": "3.17.254.50:4999/",
  "basePath": "apis",
...
  "paths": {
...
    "/check_email": {
      "post": {
        "tags": [
          "provider"
        ],
        "summary": "check email",
        "description": " check signup email exist or not",
        "operationId": "check_email",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "formData",
            "description": "email",
            "required": true,
            "type": "string"
          }
       
          
        ],
        "responses": {
          "200": {
            "description": "Email verified successfully"
          },
          "400": {
            "description": "Email Already Exist"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "APIKeyHeader": {
      "type": "apiKey",
      "in": "header",
      "name": "security_key"
    }
  },
...
  "schemes": [
    "http"
  ]
}